Refactor tests to meet DRY
