+++ jest.config.json
@@ -8,5 +8,6 @@
 	"transform": {
 		"^.+\\.(js|jsx)$": "<rootDir>/node_modules/babel-jest"
 	},
-	"transformIgnorePatterns": ["<rootDir>/node_modules/(?!@constantiner)/"]
+	"transformIgnorePatterns": ["<rootDir>/node_modules/(?!@constantiner)/"],
+	"testPathIgnorePatterns": ["<rootDir>/node_modules/", "<rootDir>/__tests__/util/"]
 }
